update script runner

+++ ogc/run.py
@@ -1,4 +1,6 @@
 import os
+import shlex
+import subprocess
 import tempfile
 from pathlib import Path
 
@@ -11,35 +13,26 @@ def make_executable(path):
     os.chmod(str(path), mode)
 
 
-def script(script_data, env, log, timeout=None, concurrent=False):
+def script(script_data, env, **kwargs):
     # preserve color
     # script --flush \
     #        --quiet \
     #        --return /tmp/ansible-output.txt \
     #        --command "my-ansible-command"
-    _run = sh.env
-    if "sudo" in script_data:
-        _run = sh.contrib.sudo.env
-    if not script_data[:2] != "#!":
-        script_data = "#!/bin/bash\n" + script_data
-    tmp_script = tempfile.mkstemp()
-    tmp_script_path = Path(tmp_script[-1])
-    tmp_script_path.write_text(script_data, encoding="utf8")
-    make_executable(tmp_script_path)
-    os.close(tmp_script[0])
-    if concurrent:
-        cmd = _run(
-            str(tmp_script_path), _env=env.copy(), _timeout=timeout, _bg=concurrent
-        )
-        cmd.wait()
+    is_single_command = len(script_data.splitlines()) == 1
+
+    args = []
+    if not is_single_command:
+        if not script_data[:2] != "#!":
+            script_data = "#!/bin/bash\n" + script_data
+        tmp_script = tempfile.mkstemp()
+        tmp_script_path = Path(tmp_script[-1])
+        tmp_script_path.write_text(script_data, encoding="utf8")
+        make_executable(tmp_script_path)
+        os.close(tmp_script[0])
+        args = ["bash", str(tmp_script_path)]
+        subprocess.run(args, env=env.copy(), **kwargs)
+        sh.rm("-rf", tmp_script_path)
     else:
-        for line in _run(
-            str(tmp_script_path),
-            _env=env.copy(),
-            _timeout=timeout,
-            _iter=True,
-            _bg_exc=False,
-            _tty_in=True,
-        ):
-            log.info(line.strip())
-    sh.rm("-rf", tmp_script_path)
+        args = shlex.split(script_data.strip())
+        subprocess.run(args, env=env.copy(), **kwargs)
